Added button on sidebar to give user choice  to select text-to-speech

+++ app.py
@@ -120,6 +120,7 @@ with st.sidebar:
     # Language selection
     user_language = st.selectbox("Select your preferred language", list(languages.values()))
     language_code = list(languages.keys())[list(languages.values()).index(user_language)]
+    text_to_speech_choice = st.selectbox("Activate text to speech?", list(["Yes", "No"]))
     
     greeting = translator.translate("Hi there! Welcome to the Center for Nonprofit Management's Resource Chatbot. How can I assist you today?", dest=language_code).text
     st.session_state.chat_history = [AIMessage(content=greeting)]
@@ -149,11 +150,12 @@ for message in st.session_state.chat_history:
         with st.chat_message("AI"):
             st.write(message.content)
             # Add a button to play the audio
-            translated_button = translator.translate("Play Audio", dest=language_code).text
-            if st.button(translated_button, key=f"play_{message.content[:10]}"):
-                audio_file = text_to_speech(message.content)
-                st.audio(audio_file, format="audio/mp3")
-                os.remove(audio_file)  # Clean up temporary file after playback
+            if text_to_speech_choice == "Yes":
+                translated_button = translator.translate("Play Audio", dest=language_code).text
+                if st.button(translated_button, key=f"play_{message.content[:10]}"):
+                    audio_file = text_to_speech(message.content)
+                    st.audio(audio_file, format="audio/mp3")
+                    os.remove(audio_file)  # Clean up temporary file after playback
     elif isinstance(message, HumanMessage):
         with st.chat_message("Human"):
             st.write(message.content)
